Time: 0 ms (100%), Space: 103.6 MB (96.89%) - LeetHub

+++ 3689-maximum-total-subarray-value-i/3689-maximum-total-subarray-value-i.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long maxTotalValue(vector<int>& nums, long long k) {
+    long long maxTotalValue(vector<int>& nums,int k) {
         long long mx = nums[0];
         long long mn = nums[0];
         for(long long x : nums){
